Factor a DaemonStore out of RemoteStore
